smile_friend_engine/requirements.txt
fastapi>=0.110.0
httpx>=0.27.0
uvicorn[standard]>=0.27.0
google-cloud-firestore>=2.16.0
